+++ frontEnd/lib/groups_home.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 import 'package:flutter/material.dart';
-import 'package:frontEnd/create_group.dart';
+import 'package:frontEnd/groups_create.dart';
 import 'package:frontEnd/imports/groups_manager.dart';
 import 'package:frontEnd/imports/users_manager.dart';
 import 'package:frontEnd/widgets/groups_list.dart';
